+++ src/com/brbr/brick/core/Component.java
@@ -1,8 +1,7 @@
-package com.brbr.core;
+package com.brbr.brick.core;
 
 import com.brbr.brick.object.GameObject;
-import com.brbr.debug.Debugger;
-import com.brbr.math.Transform;
+import com.brbr.brick.math.Transform;
 
 public class Component {
     public GameObject gameObject;
