+++ src/com/brbr/brick/math/MathExtention.java
@@ -1,4 +1,4 @@
-package com.brbr.math;
+package com.brbr.brick.math;
 
 public class MathExtention {
     public static double deg2rad(double degree){
