+++ src/com/brbr/brick/physics/ColliderType.java
@@ -1,4 +1,4 @@
-package com.brbr.physics;
+package com.brbr.brick.physics;
 
 public enum ColliderType {
     KINEMATIC,
