+++ src/com/brbr/brick/physics/PhysicManager.java
@@ -1,11 +1,8 @@
-package com.brbr.physics;
+package com.brbr.brick.physics;
 
-import com.brbr.debug.Debugger;
-import com.brbr.math.Bounds;
-import com.brbr.math.Transform;
-import com.brbr.math.Vector2;
+import com.brbr.brick.debug.Debugger;
+import com.brbr.brick.math.Bounds;
 
-import java.awt.geom.Rectangle2D;
 import java.util.*;
 
 public class PhysicManager {
